Compiler warning(s):
   rpc/masternode.cpp: In function ‘void SendMoney(const CTxDestination&, CAmount, CWalletTx&, AvailableCoinsType)’:
   rpc/masternode.cpp:47:13: warning: unused variable ‘nFeeRequired’ [-Wunused-variable]

   rpc/masternode.cpp: In function ‘UniValue masternode(const JSONRPCRequest&)’:
   rpc/masternode.cpp:133:27: warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

   rpc/systemnode.cpp: In function ‘UniValue systemnode(const JSONRPCRequest&)’:
   rpc/systemnode.cpp:75:27: warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

Resolution:

   Remove SendMoney (nothing referencing it here?).
   Change variable type from int to size_t.

+++ src/rpc/systemnode.cpp
@@ -72,7 +72,7 @@ UniValue systemnode(const JSONRPCRequest& request)
     if (strCommand == "list")
     {
         JSONRPCRequest newParams;
-        for (int i = 1; i < request.params.size(); i++) {
+        for (size_t i = 1; i < request.params.size(); i++) {
             newParams.params.push_back(request.params[i]);
         }
         return systemnodelist(newParams);
